requirements.txt
os
json
subprocess
sys
platform
google-generativeai
typing